+++ tests/conftest.py
@@ -23,17 +23,6 @@ def glottolog(glottolog_dir):
 
 
 @pytest.fixture
-def metadatafree_dataset(tmp_path):
-    values = tmp_path / 'values.csv'
-    values.write_text("""\
-ID,Language_ID,Parameter_ID,Value
-1,abcd1235,param1,val1
-2,abcd1234,param1,val2
-3,abcd1237,param1,val1""", encoding='utf8')
-    return Dataset.from_data(values)
-
-
-@pytest.fixture
 def md_path_factory():
     def mdpath(dirname):
         return pathlib.Path(__file__).parent / dirname / '{}-metadata.json'.format(
@@ -63,3 +52,22 @@ def Dictionary():
 def Generic():
     return Dataset.from_metadata(
         pathlib.Path(__file__).parent / 'Generic' / 'Generic-metadata.json')
+
+
+@pytest.fixture
+def MetadataFreeStructureDataset(tmp_path):
+    values = tmp_path.joinpath('values.csv')
+    values.write_text("""\
+ID,Language_ID,Parameter_ID,Value
+6,abcd1237,param1,val3
+1,abcd1235,param1,val1
+2,abcd1234,param1,val2
+3,book1243,param1,val3
+5,book1243,param2,val3
+4,isol1234,param1,val4""", encoding='utf8')
+    return values
+
+
+@pytest.fixture
+def metadatafree_dataset(MetadataFreeStructureDataset):
+    return Dataset.from_data(MetadataFreeStructureDataset)